backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
# GeoDjango helpers
django-environ==0.11.2
# Auth
djangorestframework-simplejwt==5.3.1
# API schema
drf-spectacular==0.27.2
requests==2.32.3